Adding Service Level Objective constructs
This adds a few different constructs that allow us to easily create alarms and dashboards
that follow Google's [Alerting on SLOs](https://landing.google.com/sre/workbook/chapters/alerting-on-slos/).

A few caveats to note:
1. Alarms cannot be done on a period greater than 1 day. Alarms dashboards will still show the 3 day window,
but this is just to show the standard windows. The alarms themselves will use a max of 1 day
2. This implementation currently groups all alarms into a single SNS topic. We will eventually need to
group these differently so that we can have better control over the severities, ie text/call vs email
3. Latencies are given to us using percentiles. There is no way to provide a custom function during aggregation.
This means we cannot map latency SLOs into an error rate like what's described in the workbook, ex: COUNTIF(Latency >= 200ms) / SUM(Requests)
Instead, this implementation uses varying percentiles for each window to achieve the same end. Smaller % windows,
such as the "2% budget burned in 1 hour" window, will use a smaller percentile. Smaller percentiles represent
the faster end of the requests in the distribution so, in effect, this is saying "if our fastest requests exceed
the latency threshold in a 1 hour period, then there is a significant issue. Send a text/call". Long term, this
approach may not be sufficient, and it will definitely have issues for services that do not have a normal distribution,
but all other solutions require using access logs and something other than CloudWatch, which we'd like to avoid for now.
4. CloudFront metrics only give OriginLatency, which means we don't have latencies for cache hits. This would also
require analyzing access logs in something other than CloudWatch
5. This will do multi burn rate alerts, but it is not currently doing multi-window with both short and long windows for
each burn rate. This is likely something I'll add shortly, but want to get this going first.