Update version workflow
This is an attempt at simplifying the workflow for publishing new
versions. I've added a package that should generate a good enough
CHANGELOG for us. This was a bit tedious, and was done in an order
that added it after the tag was created, which was a bit confusing
when looking at the history. Making this part of the scripts means
we'll just have to make sure our PR titles are descriptive enough
for the CHANGELOG.