feat(slo): Add CustomAvailability and CustomLatency SLOs
This adds the ability to create SLO Alarms and Dashboards that are based
on custom metrics. When building SLOs for other types, we can enforce
using AWS provided namespaces and metrics for the SLIs. When a custom
metric is used, such as in the case of extracting metrics from nginx
logs, the namespaces and metric names must be given to us as part of the
SLO, ex:

```typescript
const slos = [
  {
    type: 'CustomAvailability',
    namespace: 'CustomNamespace',
    errorsMetricName: 'CustomErrorCountMetric',
    countsMetricName: 'CustomRequestCountMetric',
    title: 'My Custom Availability',
    sloThreshold: 0.99,
  },
  {
    type: 'CustomLatency',
    namespace: 'CustomNamespace',
    latencyMetricName: 'CustomLatencyMetric',
    title: 'My Custom Latency',
    sloThreshold: 0.99,
    latencyThreshold: 2000,
  },
];
```

Additionally, this adds better parsing of dashboard bodies in tests.
Testing dashboards in cdk can be a bit of a pain because it's a mix of
a Fn::Import and stringified json. Added a helper function that collapses
this all into a single stringifed json to allow us to more easily do
partial matching on dashboard bodies. Rewrote the alarms-dashboard and
performance-dashboard tests to use this new method with more precise
matching.