ci: move codecoverage into separate workflow
It appears as though there's a race condition on using the CodeCoverage action in a matrix strategy, where one of the runners reports and the others do not report because there is already a report for the commit. This may lead to the Actions runner not exiting with a proper code, and therefore the required checks to not pass.